Week2/joke-generator/requirements.txt
streamlit==1.28.0
groq==0.4.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0